poultry-ai/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2.7
pandas
numpy
scikit-learn
pyyaml
joblib
